requirements.txt
DateTime==4.2
dateutils==0.6.6
Django==1.11.10
django-heroku==0.2.0
django-tastypie==0.14.1
requests==2.18.4
schedule==0.5.0
APScheduler==3.5.1